pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pscheduler-result-archiver"
authors = [{name="Komal Thareja", email="kthare10@renci.org"}]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE*"]
classifiers = ["Programming Language :: Python :: 3",
               "Operating System :: OS Independent"]
description = "Pscheduler Result Archive"
dynamic = ["version"]

keywords = ["Swagger", "Pscheduler Result Archive"]

requires-python = '>=3.9'
dependencies = [
    "connexion[flask]==2.14.2",
    "swagger-ui-bundle==0.0.9",
    "python-dateutil>=2.8.2",
    "waitress",
    ]

[project.optional-dependencies]
test = [
    "pytest~=7.1.0",
    "pytest-cov>=2.8.1",
    "pytest-randomly>=1.2.3",
    "Flask-Testing==0.8.1",
        ]

[project.urls]
Home = "https://github.com/kthare10/pscheduler-result-archiver"
Sources = "https://github.com/kthare10/pscheduler-result-archiver"

[tool.flit.module]
name = "archiver"
